Rename relative_velocity_sensor
