+++ s2e-ff/CMakeLists.txt
@@ -49,8 +49,8 @@ set(SOURCE_FILES
   src/components/aocs/initialize_relative_distance_sensor.cpp
   src/components/aocs/relative_position_sensor.cpp
   src/components/aocs/initialize_relative_position_sensor.cpp
-  src/components/aocs/RelativeVelocitySensor.cpp
-  src/components/aocs/InitializeRelativeVelocitySensor.cpp
+  src/components/aocs/relative_velocity_sensor.cpp
+  src/components/aocs/initialize_relative_velocity_sensor.cpp
   src/components/ideal/RelativeAttitudeController.cpp
   src/components/ideal/InitializeRelativeAttitudeController.cpp
   src/Library/math/DualQuaternion.cpp
