+++ s2e-ff/src/components/aocs/initialize_relative_velocity_sensor.cpp
@@ -1,4 +1,4 @@
-#include "InitializeRelativeVelocitySensor.hpp"
+#include "initialize_relative_velocity_sensor.hpp"
 
 #include <components/base/initialize_sensor.hpp>
 #include <library/initialize/initialize_file_access.hpp>
