+++ s2e-ff/src/components/aocs/initialize_relative_velocity_sensor.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "RelativeVelocitySensor.hpp"
+#include "relative_velocity_sensor.hpp"
 
 RelativeVelocitySensor InitializeRelativeVelocitySensor(ClockGenerator* clock_gen, const std::string file_name, const double compo_step_time_s,
                                                         const RelativeInformation& rel_info, const Dynamics& dynamics,
